Move `Options` to own module #31
